Bf16 support for tinychat 2.0 on GEMV and GEMM (#255)
* add dispatch macro

* add int4 to bf16 fast dequantize

* supporto bf16 in gemv

* remove gemm_cuda_old.cu

* support gemm and layernorm in bf16

* tinychat front end support bfloat16 for qwen

* support bf16 pseudo awq evaluate

* algorithm side front end support bf16

+++ awq/quantize/auto_clip.py
@@ -90,7 +90,7 @@ def apply_clip(module, clip_list):
     for name, max_val in clip_list:
         layer = get_op_by_name(module, name)
         layer.cuda()
-        max_val = max_val.to(layer.weight.device)
+        max_val = max_val.to(layer.weight.device).to(layer.weight.dtype)
         org_shape = layer.weight.shape
         layer.weight.data = layer.weight.data.reshape(*max_val.shape[:2], -1)
         layer.weight.data = torch.clamp(layer.weight.data, -max_val, max_val)
